requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.9
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
axolotl==0.0.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.10
django-autofixture==0.12.1
django-cors-headers==3.10.1
django-currentuser==0.5.3
django-elasticsearch-dsl==7.2.1
django-filter==21.1
django-model-utils==4.2.0
django-phonenumber-field==6.0.0
django-rest-elasticsearch==0.4.2
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
elasticsearch==7.16.2
elasticsearch-dsl==7.4.0
factory-boy==3.2.1
Faker==11.3.0
flower==1.0.0
frozenlist==1.2.0
h11==0.12.0
humanize==3.13.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.0
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
KDF==0.3
kombu==5.2.3
Markdown==3.3.6
MarkupSafe==2.0.1
mixer==7.2.0
multidict==5.2.0
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
phonenumbers==8.12.40
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-axolotl==0.2.3
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.8.2
python3-logstash==0.4.80
python3-openid==3.2.0
pytz==2021.3
redis==4.1.0
requests==2.27.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
sqlparse==0.4.2
text-unidecode==1.3
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
wsproto==1.0.0
yarl==1.7.2
zipp==3.7.0
zope.interface==5.4.0
